dbeaver/pro#9532 cleanup 2
