+++ webapp/packages/plugin-ai-user-profile/src/AIProfileCredentialsForm/IAIProfileCredentialsFormProps.ts
@@ -1,0 +1,13 @@
+/*
+ * CloudBeaver - Cloud Database Manager
+ * Copyright (C) 2020-2026 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0.
+ * you may not use this file except in compliance with the License.
+ */
+
+import type { IFormProps } from '@cloudbeaver/core-ui';
+
+import type { IAIProfileCredentialsFormState } from './IAIProfileCredentialsFormState.js';
+
+export interface IAIProfileCredentialsFormProps extends IFormProps<IAIProfileCredentialsFormState> {}
