+++ webapp/packages/plugin-ai-user-profile/src/components/AIProfileCredentialsPanel.tsx
@@ -11,118 +11,24 @@ import { observer } from 'mobx-react-lite';
 import {
   Button,
   ColoredContainer,
-  ConfirmationDialog,
   Container,
   Form,
-  Group,
   GroupBack,
   GroupTitle,
-  InputField,
-  SAVED_VALUE_INDICATOR,
   StatusMessage,
   Text,
-  useAutoLoad,
-  useFocus,
   useForm,
   useTranslate,
 } from '@cloudbeaver/core-blocks';
 import { useService } from '@cloudbeaver/core-di';
-import { CommonDialogService, DialogueStateResult } from '@cloudbeaver/core-dialogs';
 import { ENotificationType, NotificationService } from '@cloudbeaver/core-events';
-import { type IFormState, TabList, type TabContainerPanelComponent, TabPanelList, TabsContainer, TabsState } from '@cloudbeaver/core-ui';
+import { TabList, TabPanelList, TabsState } from '@cloudbeaver/core-ui';
 import { getFirstException } from '@cloudbeaver/core-utils';
 
 import { getAIProfileCredentialsFormPart } from '../AIProfileCredentialsForm/getAIProfileCredentialsFormPart.js';
-import type { IAIProfileCredentialsFormState } from '../AIProfileCredentialsForm/IAIProfileCredentialsFormState.js';
 import { AIProfileCredentialsPanelService } from '../AIProfileCredentialsPanelService.js';
-import { AIProfilesResource } from '@cloudbeaver/plugin-ai-profiles';
 
 const CREDENTIALS_TAB_ID = 'credentials';
-interface IAIProfileCredentialsFieldsProps {
-  formState: IFormState<IAIProfileCredentialsFormState>;
-}
-
-const tabs = new TabsContainer<IAIProfileCredentialsFieldsProps>('AI Profile Credentials');
-
-tabs.add({
-  key: CREDENTIALS_TAB_ID,
-  name: 'plugin_ai_credentials_profile',
-  panel: () => AIProfileCredentialsFields,
-});
-
-const AIProfileCredentialsFields: TabContainerPanelComponent<IAIProfileCredentialsFieldsProps> = observer(function AIProfileCredentialsFields({
-  formState,
-}) {
-  const translate = useTranslate();
-  const notificationService = useService(NotificationService);
-  const commonDialogService = useService(CommonDialogService);
-  const [tokenRef] = useFocus<HTMLInputElement>({ autofocus: true });
-  const part = getAIProfileCredentialsFormPart(formState);
-  const aiProfilesResource = useService(AIProfilesResource);
-  const state = part.state;
-  const credentialsSaved = part.credentialsSaved;
-
-  useAutoLoad(AIProfileCredentialsFields, part);
-
-  async function resetCredentials(): Promise<void> {
-    if (!state) {
-      return;
-    }
-
-    const { status } = await commonDialogService.open(ConfirmationDialog, {
-      title: translate('plugin_ai_credentials_reset_title'),
-      message: 'plugin_ai_credentials_reset_confirmation',
-      confirmActionText: 'plugin_ai_credentials_reset',
-    });
-
-    if (status !== DialogueStateResult.Resolved) {
-      return;
-    }
-
-    try {
-      await aiProfilesResource.resetCredentials(formState.state.profileId);
-      notificationService.logSuccess({
-        title: 'plugin_ai_user_profile_credentials_reset',
-        message: state.profileName,
-      });
-    } catch (exception: any) {
-      notificationService.logException(exception, 'plugin_ai_credentials_reset_failed');
-    }
-  }
-
-  return (
-    <Group className="tw:w-full" small keepSize gap>
-      <Container vertical gap>
-        <InputField value={state.profileName} disabled={formState.isDisabled} readOnly>
-          {translate('plugin_ai_credentials_profile')}
-        </InputField>
-        <InputField value={state.engineName} disabled={formState.isDisabled} readOnly>
-          {translate('plugin_ai_credentials_engine')}
-        </InputField>
-        <InputField
-          ref={tokenRef}
-          state={state}
-          type="password"
-          name="token"
-          autoComplete="new-password"
-          required={!credentialsSaved}
-          disabled={formState.isDisabled}
-          placeholder={credentialsSaved ? SAVED_VALUE_INDICATOR : undefined}
-          description={credentialsSaved ? translate('ui_processing_saved') : undefined}
-        >
-          {translate('plugin_ai_credentials_token')}
-        </InputField>
-      </Container>
-      {credentialsSaved && (
-        <div>
-          <Button type="button" variant="secondary" disabled={formState.isDisabled} onClick={resetCredentials}>
-            {translate('plugin_ai_credentials_reset')}
-          </Button>
-        </div>
-      )}
-    </Group>
-  );
-});
 
 export const AIProfileCredentialsPanel = observer(function AIProfileCredentialsPanel() {
   const translate = useTranslate();
@@ -168,7 +74,7 @@ export const AIProfileCredentialsPanel = observer(function AIProfileCredentialsP
         </GroupBack>
       </GroupTitle>
       <Form context={form} contents>
-        <TabsState container={tabs} selectedId={CREDENTIALS_TAB_ID} formState={formState}>
+        <TabsState container={credentialsPanelService.parts} selectedId={CREDENTIALS_TAB_ID} formState={formState}>
           <Container noWrap vertical>
             <Container
               className="theme-border-color-background tw:relative tw:before:content-[''] tw:before:absolute tw:before:bottom-0 tw:before:w-full tw:before:border-b-2 tw:before:border-inherit"
